# design-docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: EmSim API
  version: 1.0.0-draft
  description: |
    REST/JSON API тренажёра диспетчера ДДС (RFC-001 §5).
    Аутентификация — cookie `emsim_session` (ADR-008). Все ответы несут `X-Request-ID`
    и `Cache-Control: no-store`. Ошибки — единый конверт `Error`.
    Push — Server-Sent Events (ADR-007): `/my/stream`, `/lessons/{lessonId}/stream`; формат событий — `sse-events.schema.json`.
servers:
  - url: /api/v1
...
security:
  - cookie: []
paths:
...
  /admin/users:
    get:
      tags: [admin]
      summary: Список пользователей
      description: Отсортирован по login.
      parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/User' } }, total: { type: integer } } } } } }
    post:
      tags: [admin]
      summary: Создать пользователя
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserCreate' } } } }
      responses:
        '201': { description: Создан, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Validation' }
  /admin/users/{userId}:
    parameters: [ { $ref: '#/components/parameters/UserId' } ]
    patch:
      tags: [admin]
      summary: Изменить пользователя (ФИО, роль, служба, активность, пароль)
      description: Частичное обновление — поле, отсутствующее в теле, не меняется. service_code:null очищает поле; обучаемый 112 может быть без службы, а назначение ДДС требует совпадающий профиль. Для admin/instructor код службы запрещён. Смена пароля или active:false отзывает все сессии пользователя. Пароль, заданный администратором, для ролей из PASSWORD_FORCE_CHANGE помечается «сменить при первом входе» (ADR-038). unlock:true снимает блокировку входа.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UserPatch' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: Правка сделала бы активного admin нулём (нельзя понизить или деактивировать последнего), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '422': { $ref: '#/components/responses/Validation' }
  /admin/users/{userId}/sessions:
    parameters: [ { $ref: '#/components/parameters/UserId' } ]
    get:
      tags: [admin]
      summary: Действующие сеансы пользователя (ADR-038)
      description: Когда начат, последний раз использован и истекает, и на каком РМ. Без токенов и идентификаторов сеансов.
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/UserSession' } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [admin]
      summary: Завершить все сеансы пользователя (ADR-038)
      responses:
        '204': { description: Сеансы завершены }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/workstations:
    get:
      tags: [admin]
      summary: Рабочие места
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Workstation' } } } } }
    put:
      tags: [admin]
      summary: Заменить список рабочих мест
      description: Не удаляет строки. Каждое РМ из тела — upsert по number (id и active игнорируются на входе, РМ активируется); РМ, чей number не встретился в теле, деактивируется, а не удаляется. Дубликат number в теле — 422.
      requestBody: { required: true, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Workstation' } } } } }
      responses:
        '200': { description: Актуальный список РМ после замены (включая деактивированные), content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Workstation' } } } } }
        '422': { $ref: '#/components/responses/Validation' }
...
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: emsim_session
  parameters:
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
    UserId: { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    NotFound: { description: Не найдено, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Conflict: { description: Конфликт состояния, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Validation: { description: Ошибка валидации, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_request, validation_failed, unauthorized, forbidden, not_found, conflict, rate_limited, payload_too_large, unsupported_media_type, lesson_stopped, item_closed, stale_seq, transition_not_allowed, comment_required, call_required, call_in_progress, call_not_active, call_missed, internal_error, command_id_conflict, recording_conflict, recording_deadline_passed, stale_revision, stale_recommendation, stale_draft, has_blocking_issues, unsupported_for_editor, not_enough_scenarios, dictation_busy, dictation_unavailable, maintenance_mode, account_locked, password_change_required]
            message: { type: string }
            details: { type: object, additionalProperties: true }
        request_id: { type: string }
    Role: { type: string, enum: [admin, instructor, trainee] }
    Level: { type: string, enum: [easy, medium, hard] }
...
    User:
      type: object
      required: [id, login, full_name, role, level, active]
      properties:
        id: { type: string, format: uuid }
        login: { type: string }
        full_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        service_code: { type: string, nullable: true }
        level: { $ref: '#/components/schemas/Level', description: 'Текущий уровень обучаемого. Новый пользователь всегда easy; меняется только применением рекомендации преподавателем (срез 10), не через admin API.' }
        active: { type: boolean }
        locked_until: { type: [string, 'null'], format: date-time, description: 'ADR-038: вход заблокирован до этого момента; null, если блокировки нет' }
        credentials_change_required: { type: boolean, description: 'ADR-038: пароль задан администратором и должен быть сменён при первом входе (PASSWORD_FORCE_CHANGE). Имя намеренно без слова password: тело ответа никогда его не содержит.' }
    UserCreate:
      type: object
      required: [login, password, full_name, role]
      properties:
        login: { type: string, pattern: '^[a-z0-9._-]{3,64}$' }
        password: { type: string, minLength: 8, maxLength: 256 }
        full_name: { type: string, maxLength: 200 }
        role: { $ref: '#/components/schemas/Role' }
        service_code: { type: string }
    UserPatch:
      type: object
      properties:
        password: { type: string, minLength: 8 }
        full_name: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        service_code: { type: string, nullable: true }
        active: { type: boolean }
        unlock: { type: boolean, enum: [true], description: 'ADR-038: снять блокировку входа и обнулить серию неверных паролей. Смена пароля администратором тоже снимает блокировку.' }
    UserSession:
      type: object
      required: [created_at, last_seen_at, expires_at, workstation_number, workstation_label]
      properties:
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        workstation_number: { type: [integer, 'null'] }
        workstation_label: { type: [string, 'null'] }
    Workstation:
      type: object
      required: [number]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        number: { type: integer, minimum: 1 }
        label: { type: string }
        ip_address: { type: string, nullable: true }
        active: { type: boolean, default: true }
...